+++ src/main/java/com/kodilla/ecommercee/domain/Cart.java
@@ -1,4 +1,4 @@
-package com.kodilla.ecommercee;
+package com.kodilla.ecommercee.domain;
 
 import lombok.AccessLevel;
 import lombok.Setter;
