+++ src/main/java/com/kodilla/ecommercee/domain/Product.java
@@ -1,26 +1,27 @@
-package com.kodilla.ecommercee;
+package com.kodilla.ecommercee.domain;
 
 import lombok.AccessLevel;
 import lombok.Setter;
+
 import javax.persistence.*;
 
 @Setter(AccessLevel.PRIVATE)
 @Entity
-@Table(name = "ORDERS")
-public class Order  {
+@Table(name="PRODUCTS")
+public class Product {
 
     private Long id;
     private Cart cart;
+
+    @ManyToOne
+    @JoinColumn(name="CART_ID")
+    public Cart getCart() {
+        return cart;
+    }
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     public Long getId() {
         return id;
-    }
-
-    @OneToOne(fetch = FetchType.EAGER)
-    @JoinColumn(name="CART_ID")
-    public Cart getCart() {
-        return cart;
     }
 }
