--- src/main/java/com/kodilla/ecommercee/repository/ExampleToDelete.java
@@ -1,4 +1,0 @@
-package com.kodilla.ecommercee.repository;
-
-public class ExampleToDelete {
-}
